server/pyproject.toml
[tool.poetry]
name = "delete_shiritori"
version = "0.1.0"
description = ""
authors = ["hyksm <hiyoko3m.tech@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.65.2"
uvicorn = {extras = ["standard"], version = "^0.14.0"}
aiofiles = "^0.7.0"
python-multipart = "^0.0.5"
passlib = "^1.7.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
redis = "^3.5.3"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
beautifulsoup4 = "^4.9.3"
isort = "^5.8.0"
flake8 = "^3.9.2"
mypy = "^0.902"
black = "^21.6b0"
types-redis = "^3.5.4"
requests = "^2.25.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
